+++ alibi/datasets.py
@@ -9,14 +9,47 @@ import requests
 from requests import RequestException
 from sklearn.preprocessing import LabelEncoder
 import tarfile
-from typing import Tuple
+from typing import Tuple, Union
 import logging
+from alibi.utils.data import Bunch
 
 logger = logging.getLogger(__name__)
 
+__all__ = ['fetch_adult',
+           'fetch_imagenet',
+           'fetch_movie_sentiment']
 
+
+# deprecated functions
 def imagenet(category: str = 'Persian cat', nb_images: int = 10, target_size: tuple = (299, 299),
-             min_std: float = 10., seed: int = 42) -> Tuple[np.ndarray, np.ndarray]:
+             min_std: float = 10., seed: int = 42):
+    import warnings
+    warnings.warn('`imagenet` is deprecated and will be removed soon, use `fetch_imagenet` instead',
+                  DeprecationWarning, stacklevel=2)
+    return fetch_imagenet(category=category, nb_images=nb_images, target_size=target_size, min_std=min_std,
+                          seed=seed, return_X_y=True)
+
+
+def movie_sentiment():
+    import warnings
+    warnings.warn('`movie_sentiment` is deprecated and will be removed soon, use `fetch_movie_sentiment` instead',
+                  DeprecationWarning, stacklevel=2)
+    return fetch_movie_sentiment(return_X_y=True)
+
+
+def adult(features_drop=None):
+    if features_drop is None:
+        features_drop = ["fnlwgt", "Education-Num"]
+    import warnings
+    warnings.warn('`adult` is deprecated and will be removed soon, use `fetch_adult` instead', DeprecationWarning,
+                  stacklevel=2)
+    bunch = fetch_adult(features_drop)
+    return bunch.data, bunch.target, bunch.feature_names, bunch.category_map
+
+
+def fetch_imagenet(category: str = 'Persian cat', nb_images: int = 10, target_size: tuple = (299, 299),
+                   min_std: float = 10., seed: int = 42, return_X_y: bool = False) -> Union[Bunch, Tuple[np.ndarray,
+                                                                                                         np.ndarray]]:
     """
     Retrieve imagenet images from specified category which needs to be in the mapping dictionary.
 
@@ -36,7 +69,10 @@ def imagenet(category: str = 'Persian cat', nb_images: int = 10, target_size: tu
 
     Returns
     -------
-    List with images and the labels from imagenet.
+    Bunch
+        List with images and the labels from imagenet.
+    (data, target)
+        Tuple if ``return_X_y`` is true
     """
     mapping = {'Persian cat': 'n02123394',
                'volcano': 'n09472597',
@@ -83,16 +119,24 @@ def imagenet(category: str = 'Persian cat', nb_images: int = 10, target_size: tu
     inv_label = {v: k for k, v in label_dict.items()}
     label_idx = inv_label[category]
     labels = np.array([label_idx for _ in range(nb_images)])
-    return data, labels
+
+    if return_X_y:
+        return data, labels
+
+    target_names = [category for _ in range(nb_images)]
+    return Bunch(data=data, target=labels, target_names=target_names)
 
 
-def movie_sentiment() -> Tuple[list, list]:
+def fetch_movie_sentiment(return_X_y: bool = False) -> Union[Bunch, Tuple[list, list]]:
     """
     The movie review dataset, equally split between negative and positive reviews.
 
     Returns
     -------
-    Movie reviews and sentiment labels (0 means 'negative' and 1 means 'positive').
+    Bunch
+        Movie reviews and sentiment labels (0 means 'negative' and 1 means 'positive').
+    (data, target)
+        Tuple if ``return_X_y`` is true
     """
     url = 'http://www.cs.cornell.edu/People/pabo/movie-review-data/rt-polaritydata.tar.gz'
     try:
@@ -113,10 +157,15 @@ def movie_sentiment() -> Tuple[list, list]:
             data.append(line.decode('utf8').strip())
             labels.append(i)
     tar.close()
-    return data, labels
+    if return_X_y:
+        return data, labels
+
+    target_names = ['negative', 'positive']
+    return Bunch(data=data, target=labels, target_names=target_names)
 
 
-def adult(features_drop: list = ["fnlwgt", "Education-Num"]) -> Tuple[np.ndarray, np.ndarray, list, dict]:
+def fetch_adult(features_drop: list = None, return_X_y: bool = False) -> Union[Bunch, Tuple[np.ndarray,
+                                                                                            np.ndarray]]:
     """
     Downloads and pre-processes 'adult' dataset.
     More info: http://mlr.cs.umass.edu/ml/machine-learning-databases/adult/
@@ -124,19 +173,25 @@ def adult(features_drop: list = ["fnlwgt", "Education-Num"]) -> Tuple[np.ndarray
     Parameters
     ----------
     features_drop
-        List of features to be dropped from dataset
+        List of features to be dropped from dataset, by default drops ["fnlwgt", "Education-Num"]
 
     Returns
     -------
-    Dataset, labels, a list of features and a dictionary containing a list with the potential categories
-    for each categorical feature where the key refers to the feature column.
+    Bunch
+        Dataset, labels, a list of features and a dictionary containing a list with the potential categories
+        for each categorical feature where the key refers to the feature column.
+    (data, target)
+        Tuple if ``return_X_y`` is true
     """
+    if features_drop is None:
+        features_drop = ["fnlwgt", "Education-Num"]
+
     # download data
     dataset_url = 'http://mlr.cs.umass.edu/ml/machine-learning-databases/adult/adult.data'
     raw_features = ['Age', 'Workclass', 'fnlwgt', 'Education', 'Education-Num', 'Marital Status',
                     'Occupation', 'Relationship', 'Race', 'Sex', 'Capital Gain', 'Capital Loss',
                     'Hours per week', 'Country', 'Target']
-    raw_data = pd.read_csv(dataset_url, names=raw_features, delimiter=', ').fillna('?')
+    raw_data = pd.read_csv(dataset_url, names=raw_features, delimiter=', ', engine='python').fillna('?')
 
     # get labels, features and drop unnecessary features
     labels = (raw_data['Target'] == '>50K').astype(int).values
@@ -210,5 +265,9 @@ def adult(features_drop: list = ["fnlwgt", "Education-Num"]) -> Tuple[np.ndarray
 
     # only return data values
     data = data.values
+    target_names = ['<=50K', '>50K']
 
-    return data, labels, features, category_map
+    if return_X_y:
+        return data, labels
+
+    return Bunch(data=data, target=labels, feature_names=features, target_names=target_names, category_map=category_map)